Removing longValue method

+++ pulsar-client/src/main/java/org/apache/pulsar/client/impl/ReaderImpl.java
@@ -92,7 +92,7 @@ public class ReaderImpl<T> implements Reader<T> {
                 partitionIdx, consumerFuture, SubscriptionMode.NonDurable, readerConfiguration.getStartMessageId(), schema, null,
                 client.getConfiguration().getDefaultBackoffIntervalNanos(), client.getConfiguration().getMaxBackoffIntervalNanos());
         try {
-            if (readerConfiguration.getTimestamp().longValue() != -1) {
+            if (readerConfiguration.getTimestamp() != -1) {
                 consumer.seek(readerConfiguration.getTimestamp());
             }
         } catch (PulsarClientException exc) {
